// src/renderer/features/mon-hoc-bac-dao-tao/types.ts
export interface MonHocBacDaoTaoFilter {
  idBacDaoTao?: string;
  idKhoa?: string;
  idToBoMon?: string;
  idLoaiMonHoc?: string;
  idTinhChatMonHoc?: string;
  idKhoiKienThuc?: string;
  idHinhThucThi?: string;
  idLoaiHinhGiangDay?: string;
  idMonHoc?: string;
}

export interface AddMonHocBacDaoTaoFilter {
  idKhoaChuQuan?: string;
  idLoaiMonHoc?: string;
  monHoc?: string;
  maTuQuan?: string;
  idLoaiTietHoc?: string;
}